Fixing TEST2

+++ src/Modules/Reviews/GameRa.Modules.Reviews.Application/Reviews/GetAverageRatingByGameId/AverageRatingResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace GameRa.Modules.Reviews.Application.Reviews.GetAverageRatingByGameId;
 
-public sealed record AverageRatingResponse(Guid GameId, double AverageRating, int TotalReviews);
+public sealed record AverageRatingResponse(Guid GameId, decimal AverageRating, long TotalReviews);
